Bugfix: Address Re-Rendering Previously Hidden Posts Issue
If the extension has filtered posts as the user scrolls down, and then the user proceeds to scroll back up - previously filtered posts become un-filtered.

This is because when we initially gather post info, we don't update the element again, and this causes issues when attempting to re-filer previously seen posts.